Created a new shared module for re-used functions
